requirements.txt
Django==2.0
django-cors-headers==2.1.0
djangorestframework==3.7.7
permission==0.4.1
pytz==2017.3
